+++ internal/view/middleware.go
@@ -6,26 +6,42 @@ import (
 	"github.com/slok/grafterm/internal/view/template"
 )
 
-// withWidgetDataMiddleware will wrap a widget and every time is synced it will
-// add the static data to the sync data so it has it's corresponding
-// data so the widget itself doesn't need to store static data like the
-// dashboard, widget static data or similar.
-func withWidgetDataMiddleware(data template.Data, next widget) widget {
+// withWidgetDataMiddleware controls the variables data
+// the widget receives, it wraps any widget and will
+// mutate the variable data (updating, adding, deleting...)
+// the widget receives on every sync.
+//
+// It has the static data the widget will receive on all
+// the syncs, this way the widget doesn't need to store
+// the static data.
+//
+// It also controls the data that the user wants to override
+// (for example via cmd flags).
+//
+// Priority chain.
+// 1- OverrideData
+// 2- SyncData
+// 3- StaticData
+func withWidgetDataMiddleware(data template.Data, overrideData template.Data, next widget) widget {
 	return &widgetDataMiddleware{
-		staticData: data,
-		next:       next,
+		staticData:   data,
+		overrideData: overrideData,
+		next:         next,
 	}
 }
 
 type widgetDataMiddleware struct {
-	staticData template.Data
-	next       widget
+	staticData   template.Data
+	overrideData template.Data
+	next         widget
 }
 
 func (w widgetDataMiddleware) sync(ctx context.Context, cfg syncConfig) error {
 	// Add the sync data to the static data and place it again on the cfg.
 	data := w.staticData.WithData(cfg.templateData)
-	cfg.templateData = data
+	// Override the data asked by the user.
+	data = data.WithData(w.overrideData)
 
+	cfg.templateData = data
 	return w.next.sync(ctx, cfg)
 }